Extract advisor brief narrative mapper
