+++ quality/architecture_rules.md
@@ -35,11 +35,11 @@ report-only baseline is reviewed.
 
 The largest current modularity risks are:
 
-1. `src/app/services/advisor_brief_service.py` at 861 script-counted lines,
-2. `src/app/services/proposal_service.py` at 854 script-counted lines,
-3. `src/app/services/performance_workspace_service.py` at 842 script-counted lines,
-4. `src/app/contracts/dpm_command_center.py` at 841 script-counted lines,
-5. `src/app/contracts/proposals.py` at 812 script-counted lines.
+1. `src/app/services/proposal_service.py` at 854 script-counted lines,
+2. `src/app/services/performance_workspace_service.py` at 842 script-counted lines,
+3. `src/app/contracts/dpm_command_center.py` at 841 script-counted lines,
+4. `src/app/contracts/proposals.py` at 812 script-counted lines,
+5. `src/app/contracts/advisor_brief.py` at 812 script-counted lines.
 
 The prior longest function, `register_routers` in `src/app/router_registry.py`, has been split
 into explicit route-family groups and a short registration loop. Advise-owned router groups have
@@ -47,8 +47,9 @@ also been extracted into `src/app/router_groups/advisory.py`, reducing `router_r
 script-counted lines while preserving concrete route registration. The performance workspace
 response builder has also been split into request-context, summary/detail, evidence, and assembly
 helpers. Advisor-brief source-context, fallback narrative, source-metric, supportability, route,
-and AI fact-bundle shaping are split into `advisor_brief_source.py`, leaving the service focused on
-orchestration, AI narrative parsing, runtime supportability, and response assembly. The risk drawdown mapper has been split into
+AI fact-bundle shaping, and AI narrative parsing are split into `advisor_brief_source.py` and
+`advisor_brief_narrative.py`, leaving the service focused on orchestration, runtime supportability,
+review actions, and response assembly. The risk drawdown mapper has been split into
 period mapping, supportability, state, metadata, and payload helpers. The risk rolling mapper has
 been split into period mapping, dependency context, supportability, state, metadata, Sharpe
 fallback, and payload helpers. The risk attribution mapper has been split into period mapping,
