+++ wiki/Validation-and-CI.md
@@ -460,6 +460,16 @@ workflow action-runtime gate, mypy over 506 source files, Workbench/OpenAPI cont
 coverage tests, 94.17% total coverage, migration contract smoke, and no known vulnerabilities
 after the governed `PYSEC-2026-161` exception.
 
+The current advisor brief narrative mapper branch ratchets the source-file threshold to 854
+script-counted lines after moving AI task-request construction, AI narrative parsing, fallback
+audit normalization, and AI evidence-reference mapping into
+`src/app/services/advisor_brief_narrative.py`. It reduces
+`src/app/services/advisor_brief_service.py` from 861 to 435 script-counted lines while preserving
+advisor brief orchestration and review-action behavior. Focused validation passed with ruff check,
+ruff format check, 24 advisor-brief source/narrative/service unit tests, and refactor-threshold
+trials proving `max_source_file_lines=854` passes while `853` fails on
+`src/app/services/proposal_service.py`.
+
 The current portfolio transaction-summary context branch moves reporting-window resolution, YTD
 transaction pagination, defensive page-row extraction, reporting-currency fallback, and
 requested-window filtering into `portfolio_transaction_summary.py`. It reduces
